pushing common function
